API redesign (proceeding).
Introduce the concept of Resources, Tasks and Models. Task execution is
implemented concurrently and asynchronously by means of background
threads managed by a resource.
Why was the concept of a task necessary? Tasks basically are intended to
provide a wrapper for model execution. They provide a top-level object
for easier pickling. Also, both the task's state and model results
are recorded by themselves within a DB in order to keep resources
stateless.

The REST API provides the following routes:

  - /v1/<model>/runs
  - /v1/<model>/runs/<task_id>

Resources are implemented using Blueprints.