[ENH]: Parse and validate input parameters
From now on, a seismic catalog is received as QuakeML. SaSS exemplarily
shows how to parse the data into an obspy catalog. Note, that currently, the
catalog's spatial information is not transformed into the model's local
SRS.

Model parameters are received as a plain dict. This approach guarantees
most flexibility when additional parameters might be introduced in
future.

Reservoir related data is transmitted by means of the WKT format. The
data is validated using OGR facilities.