[BUG] Using multiprocessing spawn instead of fork
A logging queue handled in the main process seems to be required having
changed the multiprocessing module to use spawn instead of the default
fork. Passing this queue to the new process means that the queue must be
passed to each logger. Queue is not currently being passed to the model
code. This is a problem if we want to store the logs somewhere, as these
will not be written in the same way.