[BUG] Fix logging levels
When spawning processes, not everything is copied as with fork. This
means that the logging config is not copied over and this needs to be
reinitialized within the new process. Therefore a further function was
created in utils to help with this as before it was done on the app
level.
The variables that hold details about the logging config are now passed
from the app level to the task.