[ENH] Changes to Reservoir
The reservoir format and processing has been updated so that the gdal
library is not used. Instead a reservoir will be passed as json in the
request along with a spatial proj4 string and local coordinate reference
point. The output results will now reference the coordinates of the
bounding box as floats rather that a geometry type.